% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{fecha_mas_cercana}
\alias{fecha_mas_cercana}
\title{Redondea una fecha arbitraria a la fecha de corte más cercana}
\usage{
fecha_mas_cercana(fecha)
}
\arguments{
\item{fecha}{Fecha.}
}
\description{
Redondea una fecha arbitraria a la fecha de corte más cercana
}
