% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficas.R
\name{graficar_dona}
\alias{graficar_dona}
\title{Gráfica de dona para un corte}
\usage{
graficar_dona(resumen, titulo = NULL)
}
\arguments{
\item{resumen}{Salida de \code{\link[=resumen_entidad]{resumen_entidad()}} filtrada a una fecha y unidad.}

\item{titulo}{Título opcional.}
}
\value{
Objeto ggplot.
}
\description{
Gráfica de dona para un corte
}
