% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{municipios_sf}
\alias{municipios_sf}
\title{Geometría municipal simplificada}
\format{
Objeto \code{sf} con \code{cvegeo} y \code{geometry}.
}
\source{
INEGI, Marco Geoestadístico Nacional 2025.
}
\usage{
municipios_sf
}
\description{
Polígonos del MGN 2025 simplificados al 5 \% de vértices, en WGS84.
}
\keyword{datasets}
