% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequiaMX-package.R
\docType{package}
\name{sequiaMX-package}
\alias{sequiaMX}
\alias{sequiaMX-package}
\title{sequiaMX: Datos y Análisis del Monitor de Sequía en México}
\description{
Datos del Monitor de Sequía en México (CONAGUA-SMN) en formato largo, con catálogo municipal, superficie (INEGI) y población (CONAPO), y funciones para resumir la sequía por municipio, entidad federativa y organismo de cuenca. Incluye un reporte en Quarto y una aplicación Shiny.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abaltazzar/sequia_mx}
  \item Report bugs at \url{https://github.com/abaltazzar/sequia_mx/issues}
}

}
\author{
\strong{Maintainer}: Alberto Baltazar \email{abaltazar@me.com}

Authors:
\itemize{
  \item Alberto Baltazar \email{abaltazar@me.com}
}

}
\keyword{internal}
